*: move to 'psn'
