--- cmd/doc.go
@@ -1,2 +1,0 @@
-// Package cmd contains command line interfaces.
-package cmd
