+++ doc.go
@@ -1,3 +1,3 @@
-// Package psn provides utilities to programmatically inspect Linux
-// processes, sockets (ps, ss, netstat).
+// Package psn provides utilities to programmatically
+// inspect Linux processes, sockets (ps, ss, netstat).
 package psn
