--- proc/doc.go
@@ -1,2 +1,0 @@
-// Package proc inspects Linux processes.
-package proc
