+++ schema/util.go
