--- socket/doc.go
@@ -1,2 +1,0 @@
-// Package socket implements utilities to investigate sockets (Linux ss).
-package socket
