+++ status_linux.go
@@ -1,4 +1,4 @@
-package proc
+package psn
 
 import (
 	"bytes"
