+++ status_linux_test.go
@@ -1,4 +1,4 @@
-package proc
+package psn
 
 import (
 	"fmt"
