+++ util.go
@@ -1,4 +1,4 @@
-package proc
+package psn
 
 import (
 	"os"
@@ -11,7 +11,7 @@ func isInt(s string) bool {
 }
 
 func openToRead(fpath string) (*os.File, error) {
-	f, err := os.OpenFile(fpath, os.O_RDONLY, 0777)
+	f, err := os.OpenFile(fpath, os.O_RDONLY, 0444)
 	if err != nil {
 		return nil, err
 	}
